python

+++ cell04/ex02/calculator.py
@@ -1,5 +1,5 @@
-first = int(input("Give me the first number: "))
-second = int(input("Give me the second number: "))
+first = float(input("Give me the first number: "))
+second = float(input("Give me the second number: "))
 print("Thank you!")
 print(f"{first} + {second} = {first + second}")
 print(f"{first} - {second} = {first - second}")
